Move SessionExporter to `session`
